User Data Partition update for find5
Adds the ability to increase the userdata partition for
find5 (16/32GB) models. The user gets to select the data
size that he would like to keep and press a button. Rest
of the process happens automagically. User can select
the data partition between 6GB and 10GB with default being
10GB. It is assumed that
1. ADB drivers are installed and is available in the path
2. Openrecovery TWRP is available and installed
3. mke2fs, tune2fs and mkdosfs are available in sbin
Troubleshooting and extra files can be downloaded from
http://forum.xda-developers.com/showpost.php?p=57638797&postcount=34